refactor: split main logic into separate modules
